refactor : RedisRepository 인터페이스 없에고 구현체만 두기
